sqlite_to_postgres/pyproject.toml
[tool.poetry]
name = "sqlite-to-postgres"
version = "0.1.0"
description = "Загрузка данных в БД Postgres"
authors = ["aigt"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.17.0"
asyncpg = "^0.26.0"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.3"
pytest = "^7.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
